# documentation/interface-participant-openapi.json
{
  "servers": [
    {
      "url": ""
    }
  ],
  "openapi": "3.0.3",
  "info": {
    "x-logo": {
      "url": "https://bceao.int/themes/medianet/image/png/logo_fr.png"
    },
    "title": "API exposée par le participant à l'AIP",
    "version": "2.4.0",
    "description": "Le participant qui veut utiliser l'AIP devra mettre en place une API qui lui permettra de recevoir des données provenant du système interopérable.\nVeuillez vous référer à la documentation détaillée de chaque API pour plus d'informations sur les endpoints, les paramètres et les réponses attendues.\n",
    "contact": {
      "email": "pisfn@bceao.int"
    },
    "license": {
      "name": "BCEAO",
      "url": "https://www.bceao.int"
    }
  },
...
  "paths": {
    "/alias/recherche/reponses": {
      "post": {
        "tags": [
          "Alias"
        ],
        "summary": "Recevoir une réponse à une demande de recherche d'alias",
        "description": "C'est ici que seront envoyées toutes les réponses aux demandes de recherche d'alias.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReponseRechercheAlias"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "La demande de recherche d'alias a été initiée avec succès.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseSuccessRechercheAlias"
                }
              }
            }
          },
          "400": {
            "description": "Les données en entrée sont invalides.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseError"
                }
              }
            }
          }
        }
      }
    },
    "/alias/creation/reponses": {
      "post": {
        "tags": [
          "Alias"
        ],
        "summary": "Recevoir une réponse à une demande de création d'alias",
        "description": "C'est l'endpoint qui permet au participant de recevoir les réponses aux demandes de création d'alias.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReponseCreationAlias"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "La demande de création d'alias a été initiée avec succès.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseSuccessCreationAlias"
                }
              }
            }
          },
          "400": {
            "description": "Les données en entrée sont invalides.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseError"
                }
              }
            }
          }
        }
      }
    },
    "/alias/modification/reponses": {
      "post": {
        "tags": [
          "Alias"
        ],
        "summary": "Recevoir une réponse à une demande de modification d'alias",
        "description": "C'est l'endpoint qui permet au participant de recevoir les réponses aux demandes de modification d'alias.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReponseModificationAlias"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "La demande de modification d'alias a été initiée avec succès.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseSuccessModificationAlias"
                }
              }
            }
          },
          "400": {
            "description": "Les données en entrée sont invalides.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseError"
                }
              }
            }
          }
        }
      }
    },
    "/alias/suppression/reponses": {
      "post": {
        "tags": [
          "Alias"
        ],
        "summary": "Recevoir une réponse à une demande de suppression d'alias",
        "description": "C'est l'endpoint qui permet au participant de recevoir les réponses aux demandes de suppression d'alias.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReponseSuppressionAlias"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "La demande de suppression d'alias a été initiée avec succès.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseSuccessSuppressionAlias"
                }
              }
            }
          },
          "400": {
            "description": "Les données en entrée sont invalides.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseError"
                }
              }
            }
          }
        }
      }
    },
...
    "/participants/liste/reponses": {
      "post": {
        "tags": [
          "Participants"
        ],
        "summary": "Recevoir la liste des participants actualisées",
        "description": "Ce point de terminaison permet au participant de recevoir la liste actualisée des participants.\n\nLa liste des participant est envoyé sous le format d'un objet JSON converti en String contenant la nouvelle liste des participants.\n\nL'API retourne une réponse HTTP 202 si la liste des participants a été reçue avec succès,\nune réponse HTTP 400 si les données en entrée sont invalides,\net une réponse HTTP 404 si l'endpoint n'a pas pu être contacté.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListeParticipantReponse"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "La demande de la liste des participants a été reçue avec succès.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseSuccess"
                }
              }
            }
          },
          "400": {
            "description": "Les données en entrée sont invalides.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseError"
                }
              }
            }
          },
          "404": {
            "description": "L'endpoint n'a pas pu être contacté.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErrorEndpoint"
                }
              }
            }
          }
        }
      }
    },
...
    "/verifications-identites/echecs": {
      "post": {
        "tags": [
          "Echecs"
        ],
        "summary": "Recevoir un rejet de vérification d'identité",
        "description": "Ce point de terminaison permet de recevoir un rejet de vérification d'identité précédemment émise par le participant.\nLa demande et le rejet de la vérification d'identité sont liées à l'aide du msgId de la vérification d'identité.\n\nIl nécessite un objet JSON contenant les informations du rejet de la vérification d'identité.\n\nL'API retourne une réponse HTTP 202 si la demande a été reçue avec succès,\nune réponse HTTP 400 si les données en entrée sont invalides,\net une réponse HTTP 404 si l'endpoint n'a pas pu être contacté.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerificationRejet"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Le rejet de la demande de vérification d'identité a été reçue avec succès.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseSuccess"
                }
              }
            }
          },
          "400": {
            "description": "Les données en entrée sont invalides.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseError"
                }
              }
            }
          },
          "404": {
            "description": "L'endpoint n'a pas pu être contacté.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErrorEndpoint"
                }
              }
            }
          }
        }
      }
    },
...
    "/messages-iso/echec-envoi": {
      "post": {
        "tags": [
          "Echecs"
        ],
        "summary": "Recevoir une notification d'échec d'envoi de message ISO pour format invalide",
        "description": "Ce point de terminaison permet de recevoir la notification d'échec de l'envoi à PI \nd'un message précédemment soumis par le backend du participant.\nLe message soumis par le backend et la notification d'échec sont liés à l'aide du msgId\nL'API retourne une réponse HTTP 202 si la notification a été reçue avec succès,\nune réponse HTTP 400 si les données en entrée sont invalides,\net une réponse HTTP 404 si l'endpoint n'a pas pu être contacté.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EchecEnvoiMsgIso"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "La notification d'échec de l'envoi de message à PI a été reçue avec succès.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseSuccess"
                }
              }
            }
          },
          "400": {
            "description": "Les données en entrée sont invalides.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseError"
                }
              }
            }
          },
          "404": {
            "description": "L'endpoint n'a pas pu être contacté.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErrorEndpoint"
                }
              }
            }
          }
        }
      }
    },
...
    "/notifications/echecs": {
      "post": {
        "tags": [
          "Echecs"
        ],
        "summary": "Recevoir une notification d'échecs",
        "description": "Ce point de terminaison permet de recevoir une notification de rejet d'un message précédemment envoyé par le participant.\nLe message et la notification de rejet sont liées à l'aide du 'msgId' du message.\n\nIl nécessite un objet JSON contenant les informations de la notification de rejet.\n\nL'API retourne une réponse HTTP 202 si la demande a été reçue avec succès,\nune réponse HTTP 400 si les données en entrée sont invalides,\net une réponse HTTP 404 si l'endpoint n'a pas pu être contacté.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationRejet"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "La notification de rejet a été reçue avec succès.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseSuccess"
                }
              }
            }
          },
          "400": {
            "description": "Les données en entrée sont invalides.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseError"
                }
              }
            }
          },
          "404": {
            "description": "L'endpoint n'a pas pu être contacté.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErrorEndpoint"
                }
              }
            }
          }
        }
      }
    },
...
    "/message-envoi/echec-http": {
      "post": {
        "tags": [
          "Echecs"
        ],
        "summary": "Recevoir une notification d'échec d'envoi de message pour cause d'erreur 500 ou 504",
        "description": "Ce point de terminaison permet de recevoir des notification d'échec d'envoi de message pour cause d'erreur 500 ou 504.\nIl nécessite un objet JSON contenant les informations de la notification.\n\nL'API retourne une réponse HTTP 202 si la notification a été reçue avec succès,\nune réponse HTTP 400 si les données en entrée sont invalides,\net une réponse HTTP 404 si l'endpoint n'a pas pu être contacté.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationEchecEnvoi500504"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "La notification a été reçue avec succès.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseSuccess"
                }
              }
            }
          },
          "400": {
            "description": "Les données en entrée sont invalides.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseError"
                }
              }
            }
          },
          "404": {
            "description": "L'endpoint n'a pas pu être contacté.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErrorEndpoint"
                }
              }
            }
          }
        }
      }
    },
    "/message-traitement/echec": {
      "post": {
        "tags": [
          "Echecs"
        ],
        "summary": "Recevoir une notification d'échec de traitement du message par l'AIP.",
        "description": "Ce point de terminaison permet de recevoir des notification d'échec de traitement du message par l'AIP.\nIl nécessite un objet JSON contenant les informations de la notification.\n\nL'API retourne une réponse HTTP 202 si la notification a été reçue avec succès,\nune réponse HTTP 400 si les données en entrée sont invalides,\net une réponse HTTP 404 si l'endpoint n'a pas pu être contacté.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationEchecTraitement"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "La notification a été reçue avec succès.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseSuccess"
                }
              }
            }
          },
          "400": {
            "description": "Les données en entrée sont invalides.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseError"
                }
              }
            }
          },
          "404": {
            "description": "L'endpoint n'a pas pu être contacté.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErrorEndpoint"
                }
              }
            }
          }
        }
      }
    },
    "/revendications/reponses": {
      "post": {
        "tags": [
          "Revendication"
        ],
        "summary": "Recevoir une réponse à une demande de revendication d'alias",
        "description": "C'est l'endpoint qui permet au participant de recevoir les réponses aux demandes de revendication d'alias.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReponseRevendicationAlias"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "La demande de revendication d'alias a été initiée avec succès.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseSuccessRevendicationAlias"
                }
              }
            }
          },
          "400": {
            "description": "Les données en entrée sont invalides.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseError"
                }
              }
            }
          }
        }
      }
    },
    "/revendications/recuperation/reponses": {
      "post": {
        "tags": [
          "Revendication"
        ],
        "summary": "Recevoir une réponse à une recupération de revendication d'alias",
        "description": "C'est l'endpoint qui permet au participant de recevoir les réponses aux recupérations de revendication d'alias.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReponseRecuperationRevendication"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "La recupération de revendication d'alias a été initiée avec succès.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseSuccessRecuperationRevendication"
                }
              }
            }
          },
          "400": {
            "description": "Les données en entrée sont invalides.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseError"
                }
              }
            }
          }
        }
      }
    },
    "/revendications/acceptation/reponses": {
      "post": {
        "tags": [
          "Revendication"
        ],
        "summary": "Recevoir une réponse à une acceptation de revendication d'alias",
        "description": "C'est l'endpoint qui permet au participant de recevoir les réponses aux acceptations de revendication d'alias.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReponseAcceptationRevendication"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "L'acceptation de revendication d'alias a été initiée avec succès.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseSuccessAcceptationRevendication"
                }
              }
            }
          },
          "400": {
            "description": "Les données en entrée sont invalides.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseError"
                }
              }
            }
          }
        }
      }
    },
    "/revendications/rejet/reponses": {
      "post": {
        "tags": [
          "Revendication"
        ],
        "summary": "Recevoir une réponse à un rejet de revendication d'alias",
        "description": "C'est l'endpoint qui permet au participant de recevoir les réponses aux rejets de revendication d'alias.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReponseRejetRevendication"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Le rejet de revendication d'alias a été initié avec succès.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseSuccessRejetRevendication"
                }
              }
            }
          },
          "400": {
            "description": "Les données en entrée sont invalides.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ReponseCreationAlias": {
        "type": "object",
        "properties": {
          "idCreationAlias": {
            "type": "string",
            "description": "identifiant de la demande de création d’Alias à laquelle cette réponse est destinée."
          },
          "alias": {
            "type": "string",
            "description": "Valeur de l’alias. Il doit être envoyé lorsque l’alias créé est de type “SHID”."
          },
          "shid": {
            "type": "string",
            "description": "Valeur de l’alias de type SHID généré lors l’alias créé est de type MBNO."
          },
          "type": {
            "type": "string",
            "enum": [
              "SHID",
              "MBNO",
              "MCOD"
            ],
            "description": "type d'alias créé\nLes valeurs supportées par la plateforme sont:\n  -\"SHID\" Adresse de paiement \n  -\"MBNO\" Numéro de téléphone\n  -\"MCOD\" Code marchand\n"
          },
          "statut": {
            "type": "string",
            "enum": [
              "SUCCES",
              "ECHEC"
            ],
            "description": "Etat de la création d’alias"
          },
          "raisonRejet": {
            "type": "string",
            "description": "Raison du rejet de la création d'alias"
          },
          "dateCreation": {
            "type": "string",
            "format": "date-time",
            "description": "Date de création de l'alias en cas de succès"
          },
          "informationsAdditionnelles": {
            "type": "string",
            "description": "Les informations explicatives sur les raisons de l’échec de la création d’alias."
          }
        },
        "required": [
          "idCreationAlias",
          "statut"
        ],
        "example": {
          "idCreationAlias": "1",
          "alias": "+225888965212",
          "shid": "123e4567-e89b-12d3-a456-426614174000",
          "type": "MBNO",
          "statut": "SUCCES",
          "dateCreation": "2023-07-17T12:34:56.145Z"
        }
      },
      "ReponseModificationAlias": {
        "type": "object",
        "properties": {
          "alias": {
            "type": "string",
            "description": "Valeur de l’alias concernée par la modification."
          },
          "statut": {
            "type": "string",
            "enum": [
              "SUCCES",
              "ECHEC"
            ],
            "description": "Etat de la modification d’alias"
          },
          "raisonRejet": {
            "type": "string",
            "description": "Raison du rejet de la modification d’alias. En cas d’échec, il est renseigné"
          },
          "dateModification": {
            "type": "string",
            "format": "date-time",
            "description": "Date de modification de l'alias en cas de succès. En cas de succès, il est renseigné"
          },
          "informationsAdditionnelles": {
            "type": "string",
            "description": "Les informations explicatives sur les raisons de l’échec de la création d’alias."
          }
        },
        "required": [
          "alias",
          "statut"
        ],
        "example": {
          "alias": "+225888965212",
          "statut": "SUCCES",
          "dateModification": "2023-07-17T12:34:56.145Z"
        }
      },
...
      "ReponseError": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Le message d'erreur",
            "example": "Les données envoyées ne sont pas conformes"
          }
        }
      },
...
      "ReponseSuccessCreationAlias": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Le message de succès",
            "example": "Réponse de création d'alias reçue avec succés"
          }
        }
      },
...
      "ReponseSuccessRechercheAlias": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Le message de succès",
            "example": "Réponse de recherche d'alias reçue avec succés"
          }
        }
      },
...
    }
  }
}